--- backend/src/auth/jwt/jwt.guard.ts
@@ -1,5 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
-
-@Injectable()
-export class JwtGuard extends AuthGuard('jwt') {}
