+++ backend/src/exceptions/auth.exception.ts
